added utility functions and make use of them. File existance check

+++ update-15.10.sh
@@ -19,6 +19,18 @@ control_c()
 # trap keyboard interrupt (control-c)
 trap control_c SIGINT
 
+##### SETUP FUNCTIONS ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### #####
+
+export DEBIAN_FRONTEND=noninteractive
+
+function aptGetInstall {
+  apt-get install -y -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" "$@"
+}
+
+function addRepository {
+  add-apt-repository -y "$@"
+}
+
 ##### UPDATE ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### #####
 
 apt-get update
@@ -26,32 +38,37 @@ apt-get upgrade
 apt-get dist-upgrade
 
 # python-software-properties installs add-apt-repository
-apt-get install -y python-software-properties
+aptGetInstall python-software-properties
 
 ##### REPOSITORIES ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### #####
 
 # Git
-add-apt-repository -y ppa:git-core/ppa
+addRepository ppa:git-core/ppa
 
 # Java 8
-add-apt-repository -y ppa:webupd8team/java
+addRepository ppa:webupd8team/java
 
 # Chrome
-wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add -
-sh -c 'echo "deb http://dl.google.com/linux/chrome/deb/ stable main" > /etc/apt/sources.list.d/google.list'
+if [ ! -f /etc/apt/sources.list.d/google-chrome.list ]; then
+  wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add -
+  sh -c 'echo "deb http://dl.google.com/linux/chrome/deb/ stable main" > /etc/apt/sources.list.d/google-chrome.list'
+fi
 
 # Nginx
-add-apt-repository -y ppa:nginx/development
+addRepository ppa:nginx/development
 
 # MongoDB 3.2
-apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv EA312927
-echo "deb http://repo.mongodb.org/apt/debian wheezy/mongodb-org/3.2 main" | tee /etc/apt/sources.list.d/mongodb-org-3.2.list
-
+if [ ! -f /etc/apt/sources.list.d/mongodb-org-3.2.list ]; then
+  apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv EA312927
+  echo "deb http://repo.mongodb.org/apt/debian wheezy/mongodb-org/3.2 main" | tee /etc/apt/sources.list.d/mongodb-org-3.2.list
+fi
 ### SPECIFIC STUFF LUIS USES
 
 # spotify
-apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys BBEBDCB318AD50EC6865090613B00F1FD2C19886
-echo deb http://repository.spotify.com stable non-free | tee /etc/apt/sources.list.d/spotify.list
+if [ ! -f /etc/apt/sources.list.d/spotify.list ]; then
+  apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys BBEBDCB318AD50EC6865090613B00F1FD2C19886
+  echo deb http://repository.spotify.com stable non-free | tee /etc/apt/sources.list.d/spotify.list
+fi
 
 # Maria DB
 apt-key adv --recv-keys --keyserver hkp://keyserver.ubuntu.com:80 0xcbcb082a1bb943db
@@ -61,19 +78,19 @@ add-apt-repository 'deb [arch=amd64,i386] http://sfo1.mirrors.digitalocean.com/m
 # apt-key adv --keyserver keyserver.ubuntu.com --recv-keys A902DDA375E52366
 
 # Various, in specific, truecrypt
-add-apt-repository -y ppa:stefansundin/truecrypt
+addRepository ppa:stefansundin/truecrypt
 
 # Clementine
-add-apt-repository -y ppa:me-davidsansome/clementine-dev
+addRepository ppa:me-davidsansome/clementine-dev
 
 # Equalizer
-add-apt-repository -y ppa:nilarimogard/webupd8
+addRepository ppa:nilarimogard/webupd8
 
 # Libre Office
-add-apt-repository -y ppa:libreoffice/ppa
+addRepository ppa:libreoffice/ppa
 
 # Wine
-add-apt-repository -y ppa:ubuntu-wine/ppa
+addRepository ppa:ubuntu-wine/ppa
 
 ##### UPDATE REPOSITORIES ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### ##### 
 apt-get update
